requirements.txt
python-telegram-bot[job-queue]==21.9
deep-translator==1.11.4
deepl==1.20.0
python-dotenv==1.0.1
tweepy==4.14.0
numpy==2.2.1
ftfy==6.3.1
httpx==0.28.1
pysbd==0.3.4
requests==2.32.3
pillow==11.0.0
regex==2024.11.6
asyncpg==0.30.0
lxml==5.3.0
aiopg==1.4.0
torch==2.5.1
transformers==4.47.1
sentencepiece==0.2.0
openapi-python-client==0.13.4
anthropic==0.1.9
langchain==0.3.13
gpt4all==0.2.3
langchain-community==0.3.13
langchain-core==0.3.28